requirements.txt
httpx
beautifulsoup4
fake-useragent
fastapi
fastapi-cache2
pydantic-settings
uvicorn